fix: bind device-start interval as bigint for Spin Postgres
Spin maps JSON numbers to i64. INTEGER/int4 columns fail at bind time,
so POST /auth/device/start returned an empty 500. Cast the parameter as
bigint, widen the column, and tighten the Playwright auth E2E.

+++ migrations/postgres/0002_tax_profiles_and_oauth.sql
@@ -69,7 +69,7 @@ CREATE TABLE IF NOT EXISTS buwiz_server.oauth_device_codes (
     user_id UUID REFERENCES auth_users(user_id) ON DELETE CASCADE,
     session_id TEXT,
     verification_uri TEXT NOT NULL,
-    interval_seconds INTEGER NOT NULL DEFAULT 5,
+    interval_seconds BIGINT NOT NULL DEFAULT 5,
     expires_at TIMESTAMPTZ NOT NULL,
     authorized_at TIMESTAMPTZ,
     consumed_at TIMESTAMPTZ,
